Notifications Tabb && style orginizing
The notifications fragment is added to the place that formerly belonged to the chat fragment. NotRecyAdap, Notification, NotificationsViewModel, and NotificationsFragment java classes are created/implemented. Also card_view_layout and fragment_notifications xml files are created/implemented. 

Lastly chat icon is changed to a notifications icon and the colors of the fragment "add" are changed.